requirements.txt
python-dotenv==1.0.0
aiohttp==3.11.18
flask==3.1.0
flask-cors==4.0.0
pydantic==2.4.2
Flask[async]